Fix free space check for image conversion
`fetch_to_volume_format()` used to call `check_available_space()` with
the image virtual sized expressed in gigabytes, while this function
expects bytes.

This could lead Cinder to erroneously believe it had enough
space to convert an image when in reality it would fill up the
filesystem.

This patch fixes the call to `check_available_space()` and modifies the
corresponding test to actually exercise this part of the code.